+++ core-ng/src/main/java/core/framework/web/exception/ConflictException.java
@@ -5,11 +5,14 @@ import core.framework.api.web.service.ResponseStatus;
 import core.framework.log.ErrorCode;
 import core.framework.log.Severity;
 
+import java.io.Serial;
+
 /**
  * @author neo
  */
 @ResponseStatus(HTTPStatus.CONFLICT)
 public final class ConflictException extends RuntimeException implements ErrorCode {
+    @Serial
     private static final long serialVersionUID = 7787085179989898162L;
 
     private final String errorCode;