+++ core-ng/src/main/java/core/framework/web/exception/ForbiddenException.java
@@ -5,11 +5,14 @@ import core.framework.api.web.service.ResponseStatus;
 import core.framework.log.ErrorCode;
 import core.framework.log.Severity;
 
+import java.io.Serial;
+
 /**
  * @author neo
  */
 @ResponseStatus(HTTPStatus.FORBIDDEN)
 public final class ForbiddenException extends RuntimeException implements ErrorCode {
+    @Serial
     private static final long serialVersionUID = 5472429043879214361L;
 
     private final String errorCode;