+++ core-ng/src/main/java/core/framework/web/exception/NotFoundException.java
@@ -5,11 +5,14 @@ import core.framework.api.web.service.ResponseStatus;
 import core.framework.log.ErrorCode;
 import core.framework.log.Severity;
 
+import java.io.Serial;
+
 /**
  * @author neo
  */
 @ResponseStatus(HTTPStatus.NOT_FOUND)
 public final class NotFoundException extends RuntimeException implements ErrorCode {
+    @Serial
     private static final long serialVersionUID = 8663360723004690205L;
 
     private final String errorCode;